create UserLocalStore, refactor UserRepository to use new local store, fix UTs - WPB-10169
